[Refactor: #214] - 코드 수정 중
